+++ src/module/vue/components/document-img.vue
@@ -27,10 +27,11 @@ const style = computed(() => ({
 
 const $actor = inject($ActorKey, undefined)
 const $item = inject($ItemKey, undefined)
-function click() {
+function click(event: MouseEvent) {
 	// Trigger tokenizer if this is an actor portrait
 	const tokenizer = game.modules.get('vtta-tokenizer') as any
-	if (tokenizer && !$item && $actor) {
+	// TODO: test for 'shift' key
+	if (tokenizer && !event.shiftKey && !$item && $actor) {
 		return tokenizer.api.tokenizeDoc($actor)
 	}